pyproject.toml
[project]
name = "service-spark-iasworld"
version = "0.0.1"
description = """
Service for extracting tables from the CCAO system-of-record and
uploading them to the Data Department's data warehouse
"""
keywords = ["spark", "data engineering", "local government"]
authors = [
  {name = "Dan Snow", email = "daniel.snow@cookcountyil.gov"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "boto3==1.34.159",
  "joblib==1.4.2",
  "jwt==1.3.1",
  "py4j==0.10.9.7",
  "pyarrow==17.0.0",
  "pyyaml==6.0.0",
  "requests==2.32.3"
]

[tool.ruff]
line-length = 79
